working ramp, need to ramp up to higher speed to get sufficiently large BEMF amplitude for zero-cross detection

+++ 6Step_Firmware_x1/StepManager/Inc/StepManager.h
@@ -8,7 +8,7 @@ void SM_nextStep();
 uint32_t SM_getBEMFChannel();
 
 void SM_processBEMF(uint32_t bemf);
-void SM_updateDuty(uint32_t val);
+void SM_updateDuty(uint16_t val);
 
 #endif //STEP_MANAGER_H
